+++ packages/logs/src/boot/startLogs.spec.ts
@@ -109,7 +109,7 @@ describe('logs', () => {
         foo: 'bar',
         message: 'message',
         service: 'service',
-        ddtags: 'sdk_version:test,service:service',
+        mdtags: 'sdk_version:test,service:service',
         session_id: jasmine.any(String),
         session: {
           id: jasmine.any(String),