+++ packages/logs/src/domain/assembly.spec.ts
@@ -232,18 +232,18 @@ describe('startLogsAssembly', () => {
     })
   })
 
-  describe('ddtags', () => {
+  describe('mdtags', () => {
     it('should contain and format the default tags', () => {
       lifeCycle.notify(LifeCycleEventType.RAW_LOG_COLLECTED, { rawLogsEvent: DEFAULT_MESSAGE })
-      expect(serverLogs[0].ddtags).toEqual('sdk_version:test,env:test,service:service,version:1.0.0')
+      expect(serverLogs[0].mdtags).toEqual('sdk_version:test,env:test,service:service,version:1.0.0')
     })
 
     it('should append custom tags', () => {
       lifeCycle.notify(LifeCycleEventType.RAW_LOG_COLLECTED, {
         rawLogsEvent: DEFAULT_MESSAGE,
-        ddtags: ['foo:bar'],
+        mdtags: ['foo:bar'],
       })
-      expect(serverLogs[0].ddtags).toEqual('sdk_version:test,env:test,service:service,version:1.0.0,foo:bar')
+      expect(serverLogs[0].mdtags).toEqual('sdk_version:test,env:test,service:service,version:1.0.0,foo:bar')
     })
   })